requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# AI and ML
llama-cpp-python>=0.1.78
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Math and Science
sympy>=1.12.0
scipy>=1.10.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
pyautogui>=0.9.54 